requirements.txt
asgiref==3.3.4
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
dnspython==2.1.0
firebase-admin==4.5.3
gcloud==0.18.3
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
jws==0.1.3
msgpack==1.0.2
oauth2client==4.1.3
packaging==20.9
pika==1.2.0
proto-plus==1.18.1
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pymongo==3.11.3
pyparsing==2.4.7
Pyrebase4==4.4.3
python-http-client==3.3.2
python-jwt==2.0.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
sendgrid==6.7.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
uritemplate==3.0.1
whitenoise==5.2.0
Django==3.2.10
python-dotenv==0.19.2
dj-database-url==1.2.0
psycopg2-binary==2.9.5
